New Restriction added to Succeed
It is now required that the arguments to `succeed` be the original and
its immediate (possibly modified) clone. For every call to succeed there
should be a call to `partial_clone` on the original that obtained the
first form of the clone. Also added a test to vector showing proper
usage.